+++ lab9/juwe/Woda.h
@@ -8,7 +8,7 @@
 
 class Woda: public Prowiant{
 public:
-    Woda(int pojemnosc, float procenty, float masaalko):Prowiant(pojemnosc, procenty, masaalko){};
+    Woda(int pojemnosc, int procenty, float masaalko):Prowiant(pojemnosc, procenty, masaalko){};
     Woda():Prowiant(500,0,0){};
 };
 #endif //JIMP_EXERCISES_WODA_H
